requirements.txt
anthropic>=0.8.0
pillow>=10.0.0
pyautogui>=0.9.54
mss>=9.0.1
pyimgur>=0.6.0
supabase>=2.0.0
python-dotenv>=1.0.0